Pass the PaymentSettings model instance.
We need to pass the PaymentSettings model instance to use it in some helpers. Once the new settings module is permanently enabled, this model can be passed as a dependency to the appropriate helper classes. For now, we must pass it this way to avoid errors when the new settings module is disabled.